+++ pdns/dns.hh
@@ -57,7 +57,7 @@ class RCode
 {
 public:
   enum rcodes_ { NoError=0, FormErr=1, ServFail=2, NXDomain=3, NotImp=4, Refused=5, YXDomain=6, YXRRSet=7, NXRRSet=8, NotAuth=9, NotZone=10};
-  static std::string to_s(unsigned short rcode);
+  static std::string to_s(uint8_t rcode);
   static std::vector<std::string> rcodes_s;
 };
 
@@ -65,7 +65,7 @@ class ERCode
 {
 public:
   enum rcodes_ { BADVERS=16, BADSIG=16, BADKEY=17, BADTIME=18, BADMODE=19, BADNAME=20, BADALG=21, BADTRUNC=22, BADCOOKIE=23 };
-  static std::string to_s(unsigned short rcode);
+  static std::string to_s(uint8_t rcode);
 };
 
 class Opcode
